Move the runtime to a separate module
